Further refactor
Tests compile
