+++ assembly/src/test/scala/org/clulab/reach/assembly/TestAssemblySieves.scala
@@ -1,7 +1,7 @@
 package org.clulab.reach.assembly
 
 import org.scalatest.{FlatSpec, Matchers}
-import TestUtils._
+import org.clulab.reach.TestUtils._
 import org.clulab.reach.assembly.sieves.{AssemblySieve, DeduplicationSieves, PrecedenceSieves}
