+++ export/src/main/scala/org/clulab/reach/ReachCLI.scala
@@ -7,13 +7,13 @@ import scala.collection.parallel.ForkJoinTaskSupport
 import com.typesafe.config.ConfigFactory
 import com.typesafe.scalalogging.LazyLogging
 import org.apache.commons.io.{FileUtils, FilenameUtils}
-import org.clulab.assembly._
-import org.clulab.assembly.export.{AssemblyExporter, Row}
+import org.clulab.reach.assembly._
+import org.clulab.reach.assembly.export.{AssemblyExporter, Row}
 import org.clulab.odin._
-import org.clulab.reach.darpa.OutputDegrader
-import org.clulab.reach.extern.export.MentionManager
-import org.clulab.reach.extern.export.fries._
-import org.clulab.reach.extern.export.indexcards.IndexCardOutput
+import org.clulab.reach.export.OutputDegrader
+import org.clulab.reach.utils.MentionManager
+import org.clulab.reach.export.fries._
+import org.clulab.reach.export.indexcards.IndexCardOutput
 import org.clulab.reach.mentions.CorefMention
