+++ export/src/main/scala/org/clulab/reach/export/JsonOutputter.scala
@@ -1,4 +1,4 @@
-package org.clulab.reach.extern.export
+package org.clulab.reach.export
 
 import java.io._
 import java.util.Date
@@ -7,7 +7,7 @@ import org.json4s.native.Serialization
 import org.clulab.odin.Mention
 import org.clulab.reach.ReachConstants._
 import ai.lum.nxmlreader.NxmlDocument
-// import org.clulab.assembly.Assembler
+import org.clulab.reach.assembly.Assembler
 import org.clulab.reach.FriesEntry
 import org.clulab.serialization.json._
 
@@ -26,22 +26,32 @@ trait JsonOutputter {
     * Default method to be overridden by each JSON output formatter.
     */
   def toJSON (paperId:String,
-              allMentions:Seq[Mention],
-              paperPassages:Seq[FriesEntry],
-              startTime:Date,
-              endTime:Date,
-              outFilePrefix:String): String
+    allMentions:Seq[Mention],
+    paperPassages:Seq[FriesEntry],
+    startTime:Date,
+    endTime:Date,
+    outFilePrefix:String): String
 
   def toJSON(
-      paperId: String,
-      allMentions: Seq[Mention],
-      nxmldoc: NxmlDocument,
-      startTime: Date,
-      endTime: Date,
-      outFilePrefix: String
+    paperId: String,
+    allMentions: Seq[Mention],
+    nxmldoc: NxmlDocument,
+    startTime: Date,
+    endTime: Date,
+    outFilePrefix: String
   ): String = {
     toJSON(paperId, allMentions, nxmlToEntries(nxmldoc), startTime, endTime, outFilePrefix)
   }
+
+  def writeJSON(
+    paperId:String,
+    allMentions:Seq[Mention],
+    paperPassages:Seq[FriesEntry],
+    startTime:Date,
+    endTime:Date,
+    outFilePrefix:String,
+    assemblyAPI: Assembler
+  ): Unit = writeJSON(paperId, allMentions, paperPassages, startTime, endTime, outFilePrefix)
 
   /**
     * Outputs the given mentions to the given output file in some JSON-based format.
@@ -51,19 +61,19 @@ trait JsonOutputter {
     * Default method to be overridden by each JSON output formatter.
     */
   def writeJSON (paperId:String,
-                 allMentions:Seq[Mention],
-                 paperPassages:Seq[FriesEntry],
-                 startTime:Date,
-                 endTime:Date,
-                 outFilePrefix:String)
+    allMentions:Seq[Mention],
+    paperPassages:Seq[FriesEntry],
+    startTime:Date,
+    endTime:Date,
+    outFilePrefix:String)
 
   def writeJSON(
-      paperId: String,
-      allMentions: Seq[Mention],
-      nxmldoc: NxmlDocument,
-      startTime: Date,
-      endTime: Date,
-      outFilePrefix: String
+    paperId: String,
+    allMentions: Seq[Mention],
+    nxmldoc: NxmlDocument,
+    startTime: Date,
+    endTime: Date,
+    outFilePrefix: String
   ): Unit = {
     writeJSON(paperId, allMentions, nxmlToEntries(nxmldoc), startTime, endTime, outFilePrefix)
   }
